Package refactor
Refactored package names and structure:
 - core
   + data
     - server
   + di
   + extension
   + platform

 - features
   + create
   + entries
   + trackmap
