+++ app/src/main/java/com/handysparksoft/trackmap/App.kt
@@ -2,8 +2,8 @@ package com.handysparksoft.trackmap
 
 import android.app.Application
 import com.crashlytics.android.Crashlytics
-import com.handysparksoft.trackmap.di.AppComponent
-import com.handysparksoft.trackmap.di.DaggerAppComponent
+import com.handysparksoft.trackmap.core.di.AppComponent
+import com.handysparksoft.trackmap.core.di.DaggerAppComponent
 import io.fabric.sdk.android.Fabric
 
 class App : Application() {
