+++ app/src/main/java/com/handysparksoft/trackmap/core/data/server/TrackMapService.kt
@@ -1,4 +1,4 @@
-package com.handysparksoft.trackmap.data.server
+package com.handysparksoft.trackmap.core.data.server
 
 import com.handysparksoft.domain.model.TrackMap
 import retrofit2.http.Body
