+++ app/src/main/java/com/handysparksoft/trackmap/core/di/ViewModelModule.kt
@@ -1,7 +1,7 @@
-package com.handysparksoft.trackmap.di
+package com.handysparksoft.trackmap.core.di
 
-import com.handysparksoft.trackmap.ui.creation.CreateViewModelFactory
-import com.handysparksoft.trackmap.ui.currenttrackmaps.CurrentTrackMapsViewModelFactory
+import com.handysparksoft.trackmap.features.create.CreateViewModelFactory
+import com.handysparksoft.trackmap.features.entries.CurrentTrackMapsViewModelFactory
 import com.handysparksoft.usecases.GetTrackMapsUseCase
 import com.handysparksoft.usecases.SaveTrackMapUseCase
 import dagger.Module
