+++ app/src/main/java/com/handysparksoft/trackmap/core/platform/PermissionChecker.kt
@@ -1,4 +1,4 @@
-package com.handysparksoft.trackmap.ui.common
+package com.handysparksoft.trackmap.core.platform
 
 import android.app.Activity
 import android.view.View
