+++ app/src/main/java/com/handysparksoft/trackmap/features/trackmap/MainViewModel.kt
@@ -1,8 +1,8 @@
-package com.handysparksoft.trackmap.ui.main
+package com.handysparksoft.trackmap.features.trackmap
 
 import androidx.annotation.CallSuper
 import androidx.lifecycle.ViewModel
-import com.handysparksoft.trackmap.ui.common.Scope
+import com.handysparksoft.trackmap.core.platform.Scope
 
 class MainViewModel : ViewModel(), Scope by Scope.Impl() {
     init {
