+++ app/src/main/res/layout/activity_current_track_maps.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.currenttrackmaps.CurrentTrackMapsActivity">
+    tools:context=".features.entries.CurrentTrackMapsActivity">
 
     <TextView
         android:id="@+id/textview"
