+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.MainActivity">
+    tools:context=".features.trackmap.MainActivity">
 
     <fragment
         android:id="@+id/trackMapFragment"
